fix test, increase cov
