+++ test/models/BookAttrAccessible.php
@@ -10,5 +10,4 @@ class BookAttrAccessible extends Model
     public static string $table_name = 'books';
 
     public static array $attr_accessible = ['author_id'];
-    public static array $attr_protected = ['book_id'];
 }
